+++ UnitClasses/AlienMonster.cpp
@@ -1,7 +1,7 @@
 #include "AlienMonster.h"
 #include "../Game.h"
 
-AlienMonster::AlienMonster(Game* gamePtr, int health, int power, int attackCapacity)
+AlienMonster::AlienMonster(Game* gamePtr, double health, int power, int attackCapacity)
     : Unit(gamePtr, UnitType::AM, health, power, attackCapacity)
 {}
