+++ UnitClasses/EarthGunnery.cpp
@@ -1,7 +1,7 @@
 #include "EarthGunnery.h"
 #include "../Game.h"
 
-EarthGunnery::EarthGunnery(Game* gamePtr, int health, int power, int attackCapacity)
+EarthGunnery::EarthGunnery(Game* gamePtr, double health, int power, int attackCapacity)
     : Unit(gamePtr, UnitType::EG, health, power, attackCapacity)
 {}
