+++ UnitClasses/HealUnit.cpp
@@ -1,7 +1,7 @@
 #include "HealUnit.h"
 #include "../Game.h"
 
-HealUnit::HealUnit(Game* gamePtr, int health, int power, int attackCapacity)
+HealUnit::HealUnit(Game* gamePtr, double health, int power, int attackCapacity)
     : Unit(gamePtr, UnitType::EH, health, power, attackCapacity)
 {}
